# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/users": {
      "post": {
        "operationId": "AdminController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/security-settings": {
      "patch": {
        "operationId": "AdminController_updateSecuritySettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "require2FA": true,
                  "passwordMinLength": 8,
                  "sessionTimeoutMinutes": 30
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Security settings updated."
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      },
      "get": {
        "operationId": "AdminController_getSecuritySettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "require2FA": true,
                    "passwordMinLength": 8,
                    "sessionTimeoutMinutes": 30
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/general-settings": {
      "patch": {
        "operationId": "AdminController_updateGeneralSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "siteName": "My App",
                  "supportEmail": "support@example.com",
                  "maintenanceMode": false
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "General settings updated."
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      },
      "get": {
        "operationId": "AdminController_getGeneralSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "defaultCurrency": "USD",
                    "standardFeePercent": 2.5,
                    "notifyAdmins": true,
                    "notifyUsers": false
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/audit-logs": {
      "get": {
        "operationId": "AdminController_getAuditLogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "event": "Admin login",
                        "user": "admin@example.com",
                        "date": "2026-03-15T10:23:00Z"
                      },
                      {
                        "event": "Password changed",
                        "user": "jane@company.com",
                        "date": "2026-03-14T18:02:00Z"
                      }
                    ],
                    "total": 2,
                    "page": 1,
                    "pageSize": 10,
                    "pageCount": 1,
                    "success": true
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/admin/dashboard": {
      "get": {
        "operationId": "DashboardController_getDashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Admin dashboard summary",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "totalTenants": 10,
                    "totalPayments": 100,
                    "totalVolume": 50000,
                    "activeUsers": 25
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Dashboard"
        ]
      }
    },
...
  },
  "info": {
    "title": "Payment Gateway",
    "description": "API documentation for Payment Gateway",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "x-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "x-tenant-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-tenant-id"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "profileImage": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "username",
          "role"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "x-api-key": []
    },
    {
      "x-tenant-id": []
    }
  ]
}